+++ source/Plugins/Vehicle/AdaptiveSteerComponent.cpp
@@ -76,7 +76,7 @@ namespace GASS
 
 	void AdaptiveSteerComponent::OnVelocityMessage(PhysicsVelocityEventPtr message)
 	{
-		m_VehicleSpeed = message->GetLinearVelocity().FastLength();
+		m_VehicleSpeed = static_cast<float>(message->GetLinearVelocity().FastLength());
 	}
 
 	void AdaptiveSteerComponent::OnInput(InputRelayEventPtr message)
@@ -93,7 +93,7 @@ namespace GASS
 
 		if (name == "Steer")
 		{
-			float interp = (m_VehicleSpeed - m_MaxSteerAngleAtSpeed.y)/(m_MinSteerAngleAtSpeed.y - m_MaxSteerAngleAtSpeed.y); 
+			float interp = static_cast<float>((m_VehicleSpeed - m_MaxSteerAngleAtSpeed.y)/(m_MinSteerAngleAtSpeed.y - m_MaxSteerAngleAtSpeed.y));
 			if(interp < 0 )
 				interp = 0.0;
 			if(interp > 1)
@@ -101,12 +101,12 @@ namespace GASS
 			
 			
 			
-			float steer_angle = m_MinSteerAngleAtSpeed.x  + (1.0- interp)*(m_MaxSteerAngleAtSpeed.x - m_MinSteerAngleAtSpeed.x); 
-			float rad_angle = Math::Deg2Rad(steer_angle);
+			float steer_angle = static_cast<float>(m_MinSteerAngleAtSpeed.x  + (1.0- interp)*(m_MaxSteerAngleAtSpeed.x - m_MinSteerAngleAtSpeed.x));
+			float rad_angle = static_cast<float>(Math::Deg2Rad(steer_angle));
 
 			if(m_DynamicInputPower > 0)
 			{
-				m_DesiredAngle = value*(1.0-interp)* rad_angle;
+				m_DesiredAngle = value*(1.0f-interp)* rad_angle;
 				m_DesiredAngle += pow(value,m_DynamicInputPower)*(interp)* rad_angle;
 			}
 			else
