+++ source/Plugins/Vehicle/TurretComponent.cpp
@@ -158,10 +158,10 @@ namespace GASS
 	Float TurretComponent::GetAngleOnPlane(const Vec3 &plane_normal,const Vec3 &v1,const Vec3 &v2)
 	{
 		Vec3 cross = Math::Cross(v1,v2);
-		float cos_angle = Math::Dot(v1,v2);
+		float cos_angle = static_cast<float>(Math::Dot(v1,v2));
 		if(cos_angle > 1) cos_angle = 1;
 		if(cos_angle < -1) cos_angle = -1;
-		float angle = Math::Rad2Deg(acos(cos_angle));
+		float angle = static_cast<float>(Math::Rad2Deg(acos(cos_angle)));
 		if(Math::Dot(plane_normal,cross) > 0) 
 			angle *= -1;
 		return angle;
@@ -171,10 +171,10 @@ namespace GASS
 	Float TurretComponent::GetPitchAngle(const Vec3 v1,const Vec3 v2)
 	{
 		Vec3 cross = Math::Cross(v1,v2);
-		float cos_angle = Math::Dot(v1,v2);
+		float cos_angle = static_cast<float>(Math::Dot(v1,v2));
 		if(cos_angle > 1) cos_angle = 1;
 		if(cos_angle < -1) cos_angle = -1;
-		float angle = Math::Rad2Deg(acos(cos_angle));
+		float angle = static_cast<float>(Math::Rad2Deg(acos(cos_angle)));
 		if(v1.y < v2.y) 
 			angle *= -1;
 		return angle;
@@ -396,7 +396,7 @@ namespace GASS
 		m_TurnPID.setOutputLimit(m_SteerForce*3);
 		m_TurnPID.setGain(0.00003,0.0001,0.0009);
 		m_TurnPID.set(0);
-		float turn_velocity = m_TurnPID.update(angle_to_aim_dir,delta_time);
+		float turn_velocity = static_cast<float>(m_TurnPID.update(angle_to_aim_dir,delta_time));
 		//std::cout << "turn_velocity:" << turn_velocity << "\n";
